@@ -36,12 +36,12 @@
     "@reduxjs/toolkit": "^1.9.3",
     "@sendgrid/mail": "7.7.0",
     "@vercel/analytics": "0.1.11",
-    "@vercel/og": "0.4.0",
+    "@vercel/og": "0.5.0",
     "clsx": "1.2.1",
     "cross-env": "7.0.3",
     "dayjs": "^1.11.7",
     "fast-glob": "^3.2.12",
-    "framer-motion": "^10.6.0",
+    "framer-motion": "^10.10.0",
     "graphql": "^16.6.0",
     "graphql-request": "^5.2.0",
     "gray-matter": "^4.0.3",
@@ -62,32 +62,31 @@
     "remark-gfm": "^3.0.1",
     "schema-dts": "1.1.2",
     "server-only": "0.0.1",
-    "sharp": "0.31.3",
-    "sonner": "0.2.0",
+    "sharp": "0.32.0",
+    "sonner": "0.3.0",
     "superjson": "1.12.2",
-    "tailwind-merge": "1.10.0",
+    "tailwind-merge": "1.11.0",
     "usehooks-ts": "^2.9.1",
     "zod": "^3.21.4"
   },
   "devDependencies": {
     "@redux-devtools/extension": "^3.2.5",
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/line-clamp": "^0.4.2",
     "@testing-library/react": "14.0.0",
     "@total-typescript/ts-reset": "0.4.2",
     "@types/glob": "^8.1.0",
-    "@types/node": "^18.15.3",
-    "@types/react": "^18.0.28",
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.31",
     "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.55.0",
-    "@typescript-eslint/parser": "^5.55.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
     "@vitejs/plugin-react": "^3.1.0",
     "autoprefixer": "^10.4.14",
-    "concurrently": "7.6.0",
-    "cypress": "12.8.1",
-    "eslint": "^8.36.0",
+    "concurrently": "8.0.1",
+    "cypress": "12.9.0",
+    "eslint": "^8.37.0",
     "eslint-config-next": "^13.2.4",
-    "eslint-config-prettier": "^8.7.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-mdx": "2.0.5",
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.3",
@@ -95,15 +94,15 @@
     "lint-staged": "^13.2.0",
     "next-sitemap": "^4.0.6",
     "postcss": "^8.4.21",
-    "prettier": "^2.8.5",
-    "prettier-plugin-tailwindcss": "0.2.5",
+    "prettier": "^2.8.7",
+    "prettier-plugin-tailwindcss": "0.2.6",
     "redux": "^4.2.1",
     "shell-quote": "^1.8.0",
     "tailwind-scrollbar": "^3.0.0",
-    "tailwindcss": "^3.2.7",
+    "tailwindcss": "^3.3.0",
     "tailwindcss-animate": "1.0.5",
     "typescript": "5.0.2",
-    "vitest": "^0.29.3"
+    "vitest": "^0.29.8"
   },
   "engines": {
     "node": ">=16.x"
